requirements.txt
transformers
torch==2.6
torchvision==0.21
accelerate>=0.26.0
datasets==3.2.0
pillow
torchvision>=0.6.0
datasets>=1.8.0
swanlab
deepspeed==0.17.5